Little-Projects-HTML-CSS-JS
God has surely promised His grace to the humbled: that is, to those who mourn over and despair of themselves. But a man cannot be thoroughly humbled till he realizes that his salvation is utterly beyond his own powers, counsels, efforts, will and works, and depends absolutely on the will, counsel, pleasure and work of Another -- God alone.